+++ sdk/resourcemanager/astro/armastro/models.go
@@ -10,6 +10,16 @@ package armastro
 
 import "time"
 
+// AzureResourceManagerLegacyManagedServiceIdentityV4Update - Managed service identity (system assigned and/or user assigned
+// identities)
+type AzureResourceManagerLegacyManagedServiceIdentityV4Update struct {
+	// The type of managed identity assigned to this resource.
+	Type *ManagedServiceIdentityType
+
+	// The identities assigned to this resource by the user.
+	UserAssignedIdentities map[string]*UserAssignedIdentity
+}
+
 // LiftrBaseDataOrganizationProperties - Properties specific to Data Organization resource
 type LiftrBaseDataOrganizationProperties struct {
 	// REQUIRED; Marketplace details of the resource.
@@ -66,7 +76,19 @@ type LiftrBaseMarketplaceDetails struct {
 	// REQUIRED; Offer details for the marketplace that is selected by the user
 	OfferDetails *LiftrBaseOfferDetails
 
-	// REQUIRED; Azure subscription id for the the marketplace offer is purchased from
+	// Azure subscription id for the the marketplace offer is purchased from
+	SubscriptionID *string
+
+	// Marketplace subscription status
+	SubscriptionStatus *MarketplaceSubscriptionStatus
+}
+
+// LiftrBaseMarketplaceDetailsUpdate - Marketplace details for an organization
+type LiftrBaseMarketplaceDetailsUpdate struct {
+	// Offer details for the marketplace that is selected by the user
+	OfferDetails *LiftrBaseOfferDetailsUpdate
+
+	// Azure subscription id for the the marketplace offer is purchased from
 	SubscriptionID *string
 
 	// Marketplace subscription status
@@ -86,6 +108,36 @@ type LiftrBaseOfferDetails struct {
 
 	// Plan Name for the marketplace offer
 	PlanName *string
+
+	// Subscription renewal mode
+	RenewalMode *RenewalMode
+
+	// Plan Display Name for the marketplace offer
+	TermID *string
+
+	// Plan Display Name for the marketplace offer
+	TermUnit *string
+
+	// READ-ONLY; Current subscription end date and time
+	EndDate *time.Time
+}
+
+// LiftrBaseOfferDetailsUpdate - Offer details for the marketplace that is selected by the user
+type LiftrBaseOfferDetailsUpdate struct {
+	// Offer Id for the marketplace offer
+	OfferID *string
+
+	// Plan Id for the marketplace offer
+	PlanID *string
+
+	// Plan Name for the marketplace offer
+	PlanName *string
+
+	// Publisher Id for the marketplace offer
+	PublisherID *string
+
+	// Subscription renewal mode
+	RenewalMode *RenewalMode
 
 	// Plan Display Name for the marketplace offer
 	TermID *string
@@ -256,9 +308,9 @@ type OrganizationResourceListResult struct {
 // OrganizationResourceUpdate - The type used for update operations of the OrganizationResource.
 type OrganizationResourceUpdate struct {
 	// The managed service identities assigned to this resource.
-	Identity *ManagedServiceIdentity
+	Identity *AzureResourceManagerLegacyManagedServiceIdentityV4Update
 
-	// The updatable properties of the OrganizationResource.
+	// The resource-specific properties for this resource.
 	Properties *OrganizationResourceUpdateProperties
 
 	// Resource tags.
@@ -267,6 +319,9 @@ type OrganizationResourceUpdate struct {
 
 // OrganizationResourceUpdateProperties - The updatable properties of the OrganizationResource.
 type OrganizationResourceUpdateProperties struct {
+	// Marketplace details of the resource.
+	Marketplace *LiftrBaseMarketplaceDetailsUpdate
+
 	// Organization properties
 	PartnerOrganizationProperties *LiftrBaseDataPartnerOrganizationPropertiesUpdate